+++ src/app/docs/dropdown/dropdown.module.ts
@@ -1,7 +1,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { SliceKitModule } from 'ngx-slice-kit';
+import { SharedModule } from '../../shared/shared.module';
 
 import { DemoDropdownRoutingModule } from './dropdown-routing.module';
 
@@ -24,7 +24,7 @@ const moduleDeclarations = [
         CommonModule,
         DemoDropdownRoutingModule,
         FormsModule,
-        SliceKitModule
+        SharedModule,
     ],
     declarations: moduleDeclarations,
     exports: moduleDeclarations,