+++ src/app/docs/modals/modals.module.ts
@@ -1,7 +1,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { SliceKitModule } from 'ngx-slice-kit';
+import { SharedModule } from '../../shared/shared.module';
 
 import { DemoModalsRoutingModule } from './modals-routing.module';
 
@@ -32,7 +32,7 @@ const moduleDeclarations = [
     imports: [
         CommonModule,
         DemoModalsRoutingModule,
-        SliceKitModule,
+        SharedModule,
     ],
     declarations: moduleDeclarations,
     exports: moduleDeclarations,