+++ src/app/guides/guides-routing.module.ts
@@ -1,18 +1,15 @@
-import { CUSTOM_ELEMENTS_SCHEMA, NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { GuidesComponent } from './guides.component';
 // ui-kit rules and docs
-import { ColorsComponent } from './colors/colors.component';
 import { GetStartedComponent } from './get-started/get-started.component';
 import { GlyphsComponent } from './glyphs/glyphs.component';
 import { HomeComponent } from './home/home.component';
 import { I18nComponent } from './i18n/i18n.component';
 import { ImagesGridComponent } from './images-grid/images-grid.component';
-import { MdUsageComponent } from './md-usage/md-usage.component';
 import { SchematicsComponent } from './schematics/schematics.component';
 import { StyleGuideComponent } from './style-guide/style-guide.component';
-import { TextStylesComponent } from './text-styles/text-styles.component';
 import { ThemingComponent } from './theming/theming.component';
 
 const routes: Routes = [
@@ -21,15 +18,13 @@ const routes: Routes = [
         component: GuidesComponent,
         children: [
             {path: 'home', component: HomeComponent},
-            {path: 'colors', component: ColorsComponent},
+            // {path: 'colors', component: ColorsComponent},
             {path: 'get-started', component: GetStartedComponent},
             {path: 'glyphs', component: GlyphsComponent},
             {path: 'i18n', component: I18nComponent},
             {path: 'adaptive-grid', component: ImagesGridComponent},
-            {path: 'markdown', component: MdUsageComponent},
             {path: 'schematics', component: SchematicsComponent},
             {path: 'style-guide', component: StyleGuideComponent},
-            {path: 'typography', component: TextStylesComponent},
             {path: 'theming', component: ThemingComponent},
             {path: '', redirectTo: 'home', pathMatch: 'full'},
         ]