+++ src/app/guides/schematics/schematics.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
     selector: 'app-schematics',
     templateUrl: './schematics.component.html',
-    styleUrls: ['./schematics.component.scss']
+    styleUrls: ['./schematics.component.scss', '../guides.module.scss']
 })
 export class SchematicsComponent implements OnInit {