+++ src/app/guides/style-guide/style-guide.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
     selector: 'app-style-guide',
     templateUrl: './style-guide.component.html',
-    styleUrls: ['./style-guide.component.scss']
+    styleUrls: ['./style-guide.component.scss', '../guides.module.scss']
 })
 export class StyleGuideComponent implements OnInit {