+++ src/app/guides/theming/theming.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
     selector: 'app-theming',
     templateUrl: './theming.component.html',
-    styleUrls: ['./theming.component.scss']
+    styleUrls: ['./theming.component.scss', '../guides.module.scss']
 })
 export class ThemingComponent implements OnInit {