+++ src/app/shared/layout/sidenav/sidenav.component.ts
@@ -7,12 +7,35 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 })
 export class SidenavComponent implements OnInit, OnDestroy {
 
+    routes: any[] = [];
+
     constructor() {
     }
 
 
     ngOnInit(): void {
-
+        this.routes = [
+            {
+                name: 'Guides',
+                path: '/guides',
+                children: [],
+            },
+            {
+                name: 'Documents',
+                path: '/docs',
+                children: [],
+            },
+            {
+                name: 'Resources',
+                path: '/resources',
+                children: [
+                    {
+                        name: 'About',
+                        path: '/resources/about',
+                    },
+                ],
+            },
+        ];
     }
 
     ngOnDestroy(): void {